Remove the completed lines from a position before evaluating it
For some reason this is making the AI perform slightly worse (300 lines avg instead of about 400), but it is a cleaner starting point for all evaluators and I feel like the trade off is worth it.  I believe I just need to find the correct adjustments to the weights in order to get that performance back.